requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cssmin==0.2.0
Flask==1.1.1
Flask-Assets==2.0
Flask-Bcrypt==0.7.1
Flask-HTMLmin==1.5.2
Flask-Limiter==1.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-paginate==0.5.5
Flask-Paranoid==0.2.0
Flask-WTF==0.14.2
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
hashids==1.2.0
htmlmin==0.1.12
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jsmin==2.2.2
lesscpy==0.13.0
limits==1.4.1
MarkupSafe==1.1.1
ordereddict==1.1
peewee==3.13.1
Pillow==7.0.0
ply==3.11
protobuf==3.11.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rjsmin==1.1.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
webassets==2.0
Werkzeug==0.16.0
WTForms==2.2.1
xxhash==1.4.3